refactor(project-structure): organize files into directories
