--- src/main/kotlin/ecommerce/MemberResponse.kt
@@ -1,4 +1,0 @@
-package ecommerce
-
-data class MemberResponse(val id: Long,
-                          val email: String)
