--- src/main/kotlin/ecommerce/TokenResponse.kt
@@ -1,3 +1,0 @@
-package ecommerce
-
-data class TokenResponse(val accessToken: String)
