+++ src/main/kotlin/ecommerce/auth/controller/AuthController.kt
@@ -1,19 +1,25 @@
-package ecommerce
+package ecommerce.auth.controller
 
-
+import ecommerce.auth.service.AuthService
+import ecommerce.auth.exception.AuthorizationException
+import ecommerce.auth.extractor.AuthorizationExtractor
+import ecommerce.auth.extractor.BearerAuthorizationExtractor
+import ecommerce.member.dto.MemberResponse
+import ecommerce.member.dto.TokenRequest
+import ecommerce.member.dto.TokenResponse
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.validation.Valid
 import jakarta.validation.ValidationException
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
+import org.springframework.validation.FieldError
+import org.springframework.web.bind.MethodArgumentNotValidException
 import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import org.springframework.validation.FieldError
-import org.springframework.web.bind.MethodArgumentNotValidException
 
 @RestController
 @RequestMapping("/api/members")
