+++ src/main/kotlin/ecommerce/auth/exception/AuthorizationException.kt
@@ -1,4 +1,4 @@
-package ecommerce
+package ecommerce.auth.exception
 
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ResponseStatus
