+++ src/main/kotlin/ecommerce/member/domain/Member.kt
@@ -1,4 +1,4 @@
-package ecommerce
+package ecommerce.member.domain
 
 import jakarta.validation.constraints.Email
 import jakarta.validation.constraints.NotBlank
