+++ src/main/kotlin/ecommerce/member/dto/MemberResponse.kt
@@ -1,0 +1,4 @@
+package ecommerce.member.dto
+
+data class MemberResponse(val id: Long,
+                          val email: String)
