+++ src/main/kotlin/ecommerce/member/dto/TokenRequest.kt
@@ -1,4 +1,4 @@
-package ecommerce
+package ecommerce.member.dto
 
 import jakarta.validation.constraints.Email
 import jakarta.validation.constraints.NotBlank
