+++ src/main/kotlin/ecommerce/member/dto/TokenResponse.kt
@@ -1,0 +1,3 @@
+package ecommerce.member.dto
+
+data class TokenResponse(val accessToken: String)
