+++ src/main/kotlin/ecommerce/product/controller/ProductViewController.kt
@@ -1,5 +1,7 @@
-package ecommerce
+package ecommerce.product.controller
 
+import ecommerce.product.domain.Product
+import ecommerce.product.repository.ProductRepository
 import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
 import org.springframework.web.bind.annotation.GetMapping
